+++ saleor/plugins/webhook/tests/subscription_webhooks/fixtures.py
@@ -1,7 +1,7 @@
 import pytest
 
-from saleor.webhook.event_types import WebhookEventAsyncType
-from saleor.webhook.models import Webhook
+from .....webhook.event_types import WebhookEventAsyncType
+from .....webhook.models import Webhook
 
 
 @pytest.fixture