+++ CHANGELOG.md
@@ -16,7 +16,7 @@ All notable, unreleased changes to this project will be documented in this file.
 
 ### Other changes
 - Fix failing `checkoutCustomerAttach` mutation - #9401 by @IKarbowiak
-
+- Add handling Webhook payload via graphql subscriptions (#9394)  @jakubkuc
 
 # 3.1.7